backend/requirements.txt
eventlet==0.36.1
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
fpdf2==2.7.9
imbalanced-learn==0.12.3
joblib==1.4.2
numpy==1.26.4
pandas==2.2.2
Pillow==10.4.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
scikit-learn==1.5.1
shap==0.46.0
xgboost==2.1.1
